Bump to version 0.3.0
This release is comprised of a number of minor bugfixes, many new tests, and a significant refactoring that includes many name changes: the process and internal structure of functions for converting values and variables between Clojure and MarkLogic has been overhauled. Any use of functions (even if public) other than the "standard" entry points of `create-session`, `insert-element`, and `execute-module` or `execute-xquery` may be affected by this release.